REFACTOR: centralize MNE object reconstruction
