+++ mne_denoise/sspsir/core.py
@@ -41,10 +41,9 @@ from scipy.special import expit
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from .._data import extract_data_from_mne
+from .._data import extract_data_from_mne, reconstruct_mne_object
 from .._leadfield import resolve_leadfield
 from .._logging import logger, set_log_level_from_verbose
-from ..utils import reconstruct_mne_object
 
 __all__ = ["SSPSIR", "compute_sir", "compute_sspsir"]
