+++ mne_denoise/zapline/core.py
@@ -41,7 +41,7 @@ import warnings
 
 import numpy as np
 
-from .._data import extract_data_from_mne
+from .._data import extract_data_from_mne, reconstruct_mne_object
 from .._logging import set_log_level_from_verbose
 from .._spatial import (
     apply_spatial_transform,
@@ -57,7 +57,6 @@ from ..dss.utils.segmentation import CovarianceSegmenter
 from ..dss.utils.whitening import (
     map_spatial_matrices_to_sensor_space,
 )
-from ..utils import reconstruct_mne_object
 from .adaptive import (
     apply_hybrid_cleanup,
     check_artifact_presence,
